ndntlv: add Data options
fix minor compiler errors for arm cortex

update prefix to prevent overflow

update prefix to prevent overflow

ndntlv: remove unaligned memory access for cortex m0

add ifstatements to prevent failure of sprintf

Fix CVE 2018-6953, encoded tlv value no longer accepted, if it is longer than the buffer

ndntlv: remove unaligned memory access for cortex m0

fix check for error of snprintf

add comments

add comments

+++ src/ccnl-core/include/ccnl-prefix.h
@@ -192,7 +192,7 @@ ccnl_URItoComponents(char **compVector, unsigned int *compLens, char *uri);
 */
 char *
 ccnl_prefix_to_str_detailed(struct ccnl_prefix_s *pr, int ccntlv_skip, int escape_components, int call_slash,
-                            char *buf, int buflen);
+                            char *buf, size_t buflen);
 
 /**
  * @brief Transforms a Prefix into a URI, separated by '/'
